Migraciones corregidas finalmente ya se crean correctamente

+++ database/migrations/2023_05_29_165442_create_facturas_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('facturas', function (Blueprint $table) {
             $table->id('id_factura');
             $table->unsignedBigInteger('id_contrato');
-            $table->unsignedInteger('id_cobro');
+            $table->unsignedBigInteger('id_cobro');
             $table->string('xml');
             $table->string('ruta');
             $table->foreign('id_contrato')->references('id_Contrato')->on('contratos');
